requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
